# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "lb-api",
    "description": "express-prisma-mysqlite-swagger",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/post": {
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "Post 생성",
        "description": "Post 상세내역",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "authorEmail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/{id}/views": {
      "put": {
        "tags": [
          "Post"
        ],
        "summary": "Post 상세보기",
        "description": "Post 상세내역",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/{id}": {
      "delete": {
        "tags": [
          "Post"
        ],
        "summary": "Post 삭제",
        "description": "Post 상세내역",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Post 조회",
        "description": "Post 상세내역",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "유저등록하기",
        "description": "유저등록하기",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "posts": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "유저전체 정보 가져오기",
        "description": "유저전체 정보 가져오기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{id}/drafts": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "유저상세 정보 가져오기",
        "description": "유저상세 정보 가져오기",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{id}/profile": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "유저상세 프로파일 가져오기",
        "description": "유저상세 프로파일 가져오기",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "bio": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}